Tests for version control; local-only repositories; refactoring
